+++ ContestModel/src/org/icpc/tools/contest/model/ISubmission.java
@@ -81,7 +81,7 @@ public interface ISubmission extends IContestObject {
 	 *
 	 * @return the contest time
 	 */
-	int getContestTime();
+	long getContestTime();
 
 	/**
 	 * Returns the wall clock time, in ms since the epoch.
